requirements.txt
discord.py[voice]>=2.3.0
python-dotenv>=1.0.0
yt-dlp>=2024.0.0
PyNaCl>=1.5.0
